clean navbar and visibility
